Split tests into one file per endpoint/library, for readability. Same testing order.
