+++ test/mocha.opts
@@ -1,3 +1,2 @@
 --bail
---globals server
 --exit
